renamed key "nf1/nf2genotype"
"genotype" doesn't capture all of the genetic perturbations in the NF1/NF2 datasets. for example, HS01 cells have constitutively expressed shRNA targeting NF2 that results in extreme NF2 deficiency, but it is still NF2 +/+ 

renaming the key expands it's utility. could explore changing "-/-" to "deficient"in the scenario illustrated above.