revise data, experimentalData, curatedData
Created data.json and curatedData.json
moved some terms from experimentalData to data as these should apply to both curated and experimental data.

Added some terms to curatedData that would be useful for my use-cases and some others that I am aware of.

added curatedData as a resourceType in sageCommunity.

in thinking about this more, clinicalData might be worth having in it's own file...