change waterpH key to string
This changes the metadata key type to string and updates the description to specify that the value should be entered as a number on the pH scale if known, or as acidic, neutral, or basic if exact pH is not known. These changes were requested by contributors and we don't annotate with this value so it doesn't need to be controlled.