rename validBarcodeReads to validCellBarcodeReads
The AD scrnaSeq metadata template has a column for validBarcodeReads. I think we need to make this clear in the new cell hashing metadata template that this column refers to the **cell** barcode, not the **sample** barcode, so I propose changing the name of the term to avoid confusion. Open to other suggestions though! It's possible that "validBarcodeReads" is unambiguously about cell barcodes and not sample barcodes to people who routinely work with this data.