lowercase "h" for "head of caudate nucleus"
This tissue value is the only non-acronym capitalized, and yet none of the data in the AD Portal annotated with this tissue type uses a capital "h". Since all our annotations are "head of caudate nucleus", and no data with this tissue type appears in the PEC Portal, would it make sense to change the capitalization here to lowercase for consistency?