change model name
These names refer to the same model, "NSS" is the common name per the Jax website and the UCI team. Given that the study in the portal has "NSS" in the name I think it makes sense to modify here. If we could alias terms that would be a potential use case, but that has proven too burdensome in the past when it comes to keeping different schema versions in sync.